+++ model/dto/user/loginreq/sms.go
@@ -3,11 +3,11 @@ package loginreq
 type SMS struct {
 	TimeStamp string `json:"a" validate:"required,isDigit"`
 	Phone     string `json:"phone" validate:"required,isDigit"`
-	Region    int    `json:"region" validate:"required,numeric"`
+	Region    uint16 `json:"region" validate:"required,numeric"`
 	UUID      string `json:"uuid" validate:"required,isUUIDv4"`
 	Code      int    `json:"code" validate:"required,numeric"`
 }
 
 type ChooseAccount struct {
-	UID uint64 `json:"uid" validate:"required"`
+	UID uint32 `json:"uid" validate:"required"`
 }
